New-approach: CPU pre-dequant to F32, single command buffer, GPU/CPU parity fix
- CPU pre-dequant all weights to F32 at load time (Q4_0, Q8_0, Q6_K, Q4_K, Q5_K, Q8_K, Q3_K, Q2_K, Q5_0, Q5_1, Q4_1)
- Single beginBatch/endBatch per token (llama.cpp pattern, eliminates TDR)
- Fixed GEMM N/K swap for all weight matrices (K, V, gate, up, down, LM head)
- Fixed Q4_K block layout: d(fp16)@0, dmin(fp16)@2, scales_int8(12)@4, qs(128)@16
- Fixed KV cache F32 (attention, kv_cache_write, flash_attention shaders)
- Fixed RoPE angle off-by-one (seqLen -> seqLen-1)
- Added pipeline barrier between KV cache write and attention
- Fixed FFN scratch buffer aliasing (gate == inter overlap)
- Removed 8 dead shaders (dequantize, matvec_q6_k/q8_0/q4_0, embed_q6_k/q8_0, dequant_turbo, gemm_turbo)
- MaxAE: 10-24 -> 0.009733 (< 0.01 PASS threshold)
- GPU/CPU argmax match: token 25541

+++ src/host/kv_cache.cpp
@@ -12,7 +12,7 @@ KVCacheManager::KVCacheManager(VkDevice dev, VkPhysicalDevice pdev,
 }
 
 bool KVCacheManager::allocate() {
-    size_t layerBytes = static_cast<size_t>(maxSeqLen) * nKvHeads * headDim * sizeof(uint16_t);
+    size_t layerBytes = static_cast<size_t>(maxSeqLen) * nKvHeads * headDim * sizeof(float);
 
     VkPhysicalDeviceMemoryProperties memProps;
     vkGetPhysicalDeviceMemoryProperties(physicalDevice, &memProps);
@@ -113,7 +113,7 @@ void KVCacheManager::append(uint32_t layer, const void* kData, const void* vData
     if (layer >= layers.size()) return;
     auto& l = layers[layer];
 
-    size_t rowBytes = static_cast<size_t>(nKvHeads) * headDim * sizeof(uint16_t);
+    size_t rowBytes = static_cast<size_t>(nKvHeads) * headDim * sizeof(float);
     size_t kOffset = static_cast<size_t>(l.currentSeqLen) * rowBytes;
 
     VkPhysicalDeviceProperties physProps;
